Replace ContentView with StarterView as app entry
Removed ContentView.swift and updated RestlessApp.swift to use StarterView as the initial screen. Very basic reorganizing of the project file structure. Semi completed StarterView.swift which will be the starting UI view which includes the color scheme, logo (as of now) and the name. Also added startMedia image assets for starting UI.