Feat: HomeView, startview Transitions
Created homeView file that will store the UI for the main functionality of the app. Refactored StarterView to handle transition from StartView to HomeView after a delay, and removes the obsolete homePage.swift file. Added a models file to store preset models for the UI that will be used over and over again.